BIT/requirements.txt
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
filelock==3.12.4
idna==3.4
numpy==1.26.0
oscrypto==1.3.0
packaging==23.2
pandas==2.0.3
platformdirs==3.8.1
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.19.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
snowflake-connector-python==3.2.0
sortedcontainers==2.4.0
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uuid==1.30
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
filelock==3.12.4
idna==3.4
numpy==1.26.0
oscrypto==1.3.0
packaging==23.2
pandas==2.0.3
platformdirs==3.8.1
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.19.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
snowflake-connector-python==3.2.0
sortedcontainers==2.4.0
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uuid==1.30
